+++ agent/src/main/java/com/teamscale/jacoco/agent/convert/Converter.java
@@ -60,7 +60,7 @@ public class Converter {
 				wrap(logger));
 
 		try (Benchmark benchmark = new Benchmark("Generating the XML report")) {
-			generator.convert(new Dump(sessionInfo, executionDataStore), Paths.get(arguments.outputFile));
+			generator.convert(new Dump(sessionInfo, executionDataStore), Paths.get(arguments.outputFile).toFile());
 		} catch (EmptyReportException e) {
 			logger.warn("Converted report was emtpy.", e);
 		}